Make project time Long

+++ app/models/project/Project.scala
@@ -219,7 +219,7 @@ class ProjectService @Inject()(projectDAO: ProjectDAO, teamDAO: TeamDAO, userSer
       )
     }
 
-  def publicWritesWithStatus(project: Project, openTaskInstances: Int, tracingTime: Int)(
+  def publicWritesWithStatus(project: Project, openTaskInstances: Int, tracingTime: Long)(
       implicit ctx: DBAccessContext): Fox[JsObject] =
     for {
       projectJson <- publicWrites(project)
